+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Version 0.2.0 (2023/10/01)
+
+- Added discardable, disposable and description to card setting
+- Added discard action
+- Fixed state saving
+- Added deck limit check for cards
+- Added probabilities
+
 ## Version 0.1.0 (2023/10/01)
 
 - Added basic deck editing